Assert on `queriedButton` not `button` in the test (#6)
The current test was meaningless as we didn't assert on the text content of `queriedButton`.

I've also removed `addEventListener` because it doesn't actually work. The `button` instance is already deallocated by the time that test has finished, so the closure passed to `addEventListener` is released too. We need to find a way to run async tests like this, and I have a few thoughts, but this still needs to be explored.