Improve compatibility with Embedded Swift
The change adds a conditional import excluding the JavaScriptEventLoop module that's not yet fully compatible with Embedded Swift. With the `JSObject` type from the JavaScriptKit dependency, which is also now compatible with Embedded Swift, record types are now mapped to `JSObject`. As for fields marked as optional in originating IDL, these are now correctly mapped as optional in Swift.