Clean up use of closure property wrappers
This changes improves compatibility of the WebAPIKit library with Embedded Swift. Unsupported concurrency code is excluded when concurrency is not available, including the use of async sequences. Use of closure patterns is cleaned up to use getters and setters directly instead of property wrappers, which reduces binary code size when Embedded Swift is used.